+++ CHANGELOG.md
@@ -7,7 +7,8 @@ This project loosely adheres to [semantic versioning](https://semver.org/spec/v2
 ## Unreleased [v0.1.2]
 
 - Exported underlying `st.reindex()` function to allow node relabeling
-
+- Added ability to chaneg the default simplex output type via the `s_type` Callable
+- Re-added vertex/edge contraction to the API
 
 ## [v0.1.1](https://github.com/peekxc/simplextree-py/releases/tag/v0.1.1) - 2023-08-20